requirements.txt
albumentations==0.1.2
appnope==0.1.0
backcall==0.1.0
bleach==3.0.2
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
grpcio==1.15.0
imgaug==0.2.6
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pretrainedmodels==0.7.0
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.1.0
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.4.2
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.11.0
tensorboardX==1.4
tensorflow==0.12.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
